### 2024-09-27 #7298 @editorialbot
[PRE REVIEW]: High-performance estimation of Higher-Order Interactions from multivariate data
**Submitting author:** @EtienneCmb (<a href="http://orcid.org/0000-0002-7362-3247">Etienne Combrisson</a>)
**Repository:** https://github.com/brainets/hoi
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.1
