+++ paper/paper.bib
@@ -18,6 +18,7 @@
   author={Luppi, Andrea I and Mediano, Pedro AM and Rosas, Fernando E and Holland, Negin and Fryer, Tim D and O’Brien, John T and Rowe, James B and Menon, David K and Bor, Daniel and Stamatakis, Emmanuel A},
   journal={Nature Neuroscience},
   volume={25},
+  doi = {10.1038/s41593-022-01070-0},
   number={6},
   pages={771--782},
   year={2022},
@@ -52,14 +53,18 @@
 }
 
 @article{scagliarini:2023,
-  title={Gradients of O-information: Low-order descriptors of high-order dependencies},
-  author={Scagliarini, Tomas and Nuzzi, Davide and Antonacci, Yuri and Faes, Luca and Rosas, Fernando E and Marinazzo, Daniele and Stramaglia, Sebastiano},
-  journal={Physical Review Research},
-  volume={5},
-  number={1},
-  pages={013025},
-  year={2023},
-  publisher={APS}
+  title = {Gradients of O-information: Low-order descriptors of high-order dependencies},
+  author = {Scagliarini, T. and Nuzzi, D. and Antonacci, Y. and Faes, L. and Rosas, F. E. and Marinazzo, D. and Stramaglia, S.},
+  journal = {Phys. Rev. Res.},
+  volume = {5},
+  issue = {1},
+  pages = {013025},
+  numpages = {8},
+  year = {2023},
+  month = {Jan},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevResearch.5.013025},
+  url = {https://link.aps.org/doi/10.1103/PhysRevResearch.5.013025}
 }
 
 @article{battiston:2021,
@@ -69,6 +74,7 @@
   volume={17},
   number={10},
   pages={1093--1098},
+  doi= {10.1038/s41567-021-01371-4},
   year={2021},
   publisher={Nature Publishing Group UK London}
 }
@@ -85,13 +91,16 @@
 }
 
  @article{timme:2018,
-  title={A Tutorial for Information Theory in Neuroscience},
-  volume={5},
-  ISSN={2373-2822},
-  number={3},
-  journal={eNeuro},
-  author={Timme, Nicholas M. and Lapish, Christopher},
-  year={2018},
+  author = {Timme, Nicholas M. and Lapish, Christopher},
+  title = {A Tutorial for Information Theory in Neuroscience},
+  volume = {5},
+  number = {3},
+  year = {2018},
+  doi = {10.1523/ENEURO.0052-18.2018},
+  publisher = {Society for Neuroscience},
+  URL = {https://www.eneuro.org/content/5/3/ENEURO.0052-18.2018},
+  eprint = {https://www.eneuro.org/content/5/3/ENEURO.0052-18.2018.full.pdf},
+  journal = {eNeuro}
 }
 
 @article{czyz:2024,
@@ -108,6 +117,7 @@
   journal={Neurobiology of Disease},
   volume={175},
   pages={105918},
+  doi = {10.1016/j.nbd.2022.105918},
   year={2022},
   publisher={Elsevier}
 }
@@ -119,6 +129,7 @@
   volume={11},
   number={9},
   pages={734--744},
+  doi = {10.1089/brain.2020.0982},
   year={2021},
   publisher={Mary Ann Liebert, Inc., publishers 140 Huguenot Street, 3rd Floor New~…}
 }
@@ -126,11 +137,12 @@
 @article{combrisson:2024,
   title={Neural interactions in the human frontal cortex dissociate reward and punishment learning},
[…]
