### 2024-10-08 #7298 @jbytecode
@crvernon - sure, gladly!
