### 2024-10-11 #7298 @jbytecode
👋👋👋 Dear @Saran-nns and @pitmonticone 👋👋👋

**[High-performance estimation of Higher-Order Interactions from multivariate data](https://github.com/openjournals/joss-reviews/issues/7298)**

You can find more information at the top of this Github issue ([#7298](https://github.com/openjournals/joss-reviews/issues/7298)).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be about 25 check items for each single reviewer.

Thank you in advance!
