### 2024-10-11 #7298 @pitmonticone
Dear @jbytecode,

Thank you for considering me for the review. I am happy to confirm that I am available to assist in reviewing this submission.

Additionally, I would like to suggest @ClaudMor to join the review process if available.
