### 2024-10-11 #7298 @ClaudMor
Hi @pitmonticone and @jbytecode!

I confirm I would be available for this review.
