### 2024-10-16 #7360 @jbytecode
Dear @pitmonticone, @ClaudMor

You can start your review with creating your task lists. In that list, there are several tasks.

Whenever you perform a task, you can check on the corresponding checkbox. Since the review process of JOSS is interactive, you can always interact with the author, the other reviewers, and the editor during the process. You can open issues and send pull requests in the target repo. Please mention the url of this page in there so we can keep tracking what is going on out of our world.

Please create your tasklist by typing

`@editorialbot generate my checklist`

Thank you in advance.
