### 2024-10-16 brainets/hoi#60 @pitmonticone
[REVIEW]: Missing DOIs
Dear authors,

Please review the [missing DOIs](https://github.com/openjournals/joss-reviews/issues/7360#issuecomment-2415799470) and ensure the references are updated accordingly.

Once you have completed this, please confirm by posting the following comment in the review thread [here](https://github.com/openjournals/joss-reviews/issues/7360):

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7360
