### 2024-10-16 cb3dcaf
[REVIEW]: Fix typos in docstrings
