+++ hoi/metrics/dtc.py
@@ -108,7 +108,7 @@ class DTC(HOIEstimator):
         Returns
         -------
         hoi : array_like
-            The NumPy array containing values of higher-rder interactions of
+            The NumPy array containing values of higher-order interactions of
             shape (n_multiplets, n_variables)
         """
         # ________________________________ I/O ________________________________
