+++ hoi/metrics/phiid_red.py
@@ -108,7 +108,7 @@ class RedundancyphiID(HOIEstimator):
         Returns
         -------
         hoi : array_like
-            The NumPy array containing values of higher-rder interactions of
+            The NumPy array containing values of higher-order interactions of
             shape (n_multiplets, n_variables)
         """
         # ________________________________ I/O ________________________________
