+++ hoi/metrics/sinfo.py
@@ -109,7 +109,7 @@ class Sinfo(HOIEstimator):
         Returns
         -------
         hoi : array_like
-            The NumPy array containing values of higher-rder interactions of
+            The NumPy array containing values of higher-order interactions of
             shape (n_multiplets, n_variables)
         """
         # ________________________________ I/O ________________________________
