+++ hoi/metrics/infotopo.py
@@ -123,7 +123,7 @@ class InfoTopo(HOIEstimator):
         Returns
         -------
         hoi : array_like
-            The NumPy array containing values of higher-rder interactions of
+            The NumPy array containing values of higher-order interactions of
             shape (n_multiplets, n_variables)
         """
         # ____________________________ ENTROPIES ______________________________
