+++ hoi/metrics/info_tot.py
@@ -81,7 +81,7 @@ class InfoTot(HOIEstimator):
         Returns
         -------
         hoi : array_like
-            The NumPy array containing values of higher-rder interactions of
+            The NumPy array containing values of higher-order interactions of
             shape (n_multiplets, n_variables)
         """
         # ________________________________ I/O ________________________________
