+++ hoi/metrics/rsi.py
@@ -104,7 +104,7 @@ class RSI(HOIEstimator):
         Returns
         -------
         hoi : array_like
-            The NumPy array containing values of higher-rder interactions of
+            The NumPy array containing values of higher-order interactions of
             shape (n_multiplets, n_variables)
         """
         # ________________________________ I/O ________________________________
