### 2024-10-17 brainets/hoi#65 @pitmonticone
[REVIEW]: TypeError in `plot_infotopo.py`
When running the example script `examples/metrics/plot_infotopo.py` I encountered the following error:

```console
TypeError: unsupported type for timedelta seconds component: jaxlib.xla_extension.ArrayImpl
```

REVIEW ISSUE: REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7360
