### 2024-10-17 brainets/hoi#67 @pitmonticone
[REVIEW]: TypeError in `plot_inspect_results.py`
When running the example script `examples/miscellaneous/plot_inspect_results.py` I encountered the following error:

```console
TypeError: unsupported type for timedelta seconds component: jaxlib.xla_extension.ArrayImpl
```

REVIEW ISSUE: REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7360
