### 2024-10-17 brainets/hoi#68 @pitmonticone
[REVIEW]: No plots in `plot_jax.py`
Dear authors,

The name of the example script `examples/miscellaneous/plot_jax.py` does not perform any plotting. It may be helpful to either rename the script to better reflect its functionality or modify the script to include plotting if that was the intended purpose.

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7360
