### 2024-10-17 brainets/hoi#70 @pitmonticone
[REVIEW]: SyntaxWarning in `plot_ml_vs_it.py` at lines 188 and 192
When running the example script `examples/tutorials/plot_ml_vs_it.py` I encountered the following warning at lines 188 and 192:

```console
SyntaxWarning: invalid escape sequence '\p` plt.ylabel("$\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
```

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7360
