### 2024-10-17 b76b3b1
Update plot_sim_red_syn.py
