+++ examples/tutorials/plot_sim_red_syn.py
@@ -343,3 +343,4 @@ plt.ylim(-h_max, h_max)
 plt.xlabel("Times")
 plt.ylabel("Oinfo [bits]")
 plt.title("Dynamic HOI", fontweight="bold")
+plt.show()
