### 2024-10-22 brainets/hoi#72 @ClaudMor
[REVIEW]: `get_nbest_mult` errors in Quickstart
### Description of the problem

The code in the [Quickstart](https://brainets.github.io/hoi/quickstart.html) section of the documentation returns an error.

The code in the [Quickstart](https://brainets.github.io/hoi/quickstart.html) section of the documentation:

```python
from hoi.simulation import simulate_hoi_gauss

data = simulate_hoi_gauss(n_samples=1000, triplet_character='synergy')

# import the O-information
from hoi.metrics import Oinfo

# define the model
model = Oinfo(data)

# compute hoi for multiplets with a minimum size of 3 and maximum size of 3
# using the Gaussian Copula entropy
hoi = model.fit(minsize=3, maxsize=3, method="gc")

from hoi.plot import plot_landscape
from hoi.utils import get_nbest_mult

# plot the landscape
plot_landscape(hoi, model=model)

# print the summary table
print(get_nbest_mult(hoi, model=model))
```

Errors with:

```nothing
Traceback (most recent call last):
  File "<python-input-19>", line 1, in <module>
    print(get_nbest_mult(hoi, model=model))
          ~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^
  File "C:\Users\claud\anaconda3\envs\hoi\Lib\site-packages\hoi\utils\stats.py", line 266, in get_nbest_mult
    indices = np.arange(hoi.shape[0])
                        ~~~~~~~~~^^^
IndexError: tuple index out of range
```

This issue is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/7360).

### Steps to reproduce

```Python
from hoi.simulation import simulate_hoi_gauss

data = simulate_hoi_gauss(n_samples=1000, triplet_character='synergy')

# import the O-information
from hoi.metrics import Oinfo

# define the model
model = Oinfo(data)

# compute hoi for multiplets with a minimum size of 3 and maximum size of 3
# using the Gaussian Copula entropy
hoi = model.fit(minsize=3, maxsize=3, method="gc")

from hoi.plot import plot_landscape
from hoi.utils import get_nbest_mult

# plot the landscape
plot_landscape(hoi, model=model)

# print the summary table
print(get_nbest_mult(hoi, model=model))
```

### Expected results

A summary table

### Actual results

```nothing
Traceback (most recent call last):
  File "<python-input-19>", line 1, in <module>
    print(get_nbest_mult(hoi, model=model))
          ~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^
  File "C:\Users\claud\anaconda3\envs\hoi\Lib\site-packages\hoi\utils\stats.py", line 266, in get_nbest_mult
    indices = np.arange(hoi.shape[0])
                        ~~~~~~~~~^^^
IndexError: tuple index out of range
```

### Additional information

/
