### 2024-10-23 brainets/hoi#73 @pitmonticone
[REVIEW]: Feedback on the paper
Dear authors,

Overall, I believe that the paper is well-written and satisfies most of the criteria in the JOSS checklist:

- [x] **Summary**: No revision required.
- [x]  **Statement of need**: No revision required.
- [x] **State of the field**: While the comparisons to other packages are informative, I believe this section could be improved by incorporating a "Related packages" section. Specifically, this could provide a clearer and more structured comparison of `HOI` with all the packages mentioned on the [HOI quickstart page](https://brainets.github.io/hoi/quickstart.html#other-softwares-for-the-analysis-of-higher-order-interactions). This would give readers a more comprehensive understanding of the existing tools and how `HOI` distinguishes itself.
- [x] **Quality of writing**: No revision required.
- [x]  **References**: Please consider resolving the issue #60.

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7360
