### 2024-10-27 #7360 @jbytecode
@pitmonticone - Thank you for your review, opened issues, and pull requests.

@EtienneCmb - Could you please consider the review items and apply them? Please ping us here when you've done with them. Thank you in advance!
