### 2024-10-29 cfafda5
Add DOI of Shannon's book
