+++ paper/paper.bib
@@ -85,6 +85,7 @@
   journal={The Bell system technical journal},
   volume={27},
   number={3},
+  doi={10.1002/j.1538-7305.1948.tb01338.x},
   pages={379--423},
   year={1948},
   publisher={Nokia Bell Labs}
