### 2024-10-29 brainets/hoi#60 @EtienneCmb
Hi @pitmonticone,

Missing DOI were added witht eh commit [cfafda5](https://github.com/brainets/hoi/commit/cfafda54d3127a7850e74d39bdd519df31cc7f89). Please see the report of the editorial bot [here](https://github.com/openjournals/joss-reviews/issues/7360#issuecomment-2443861769)
