### 2024-10-29 38cc6bc
Use jax-tqdm instead of custom progressbar
