+++ hoi/utils/progressbar.py
@@ -1,9 +1,4 @@
-import typing
-
-import jax
-from jax.experimental import host_callback
 from tqdm.auto import tqdm
-
 from hoi.utils.logging import logger
 
 
@@ -18,174 +13,3 @@ def get_pbar(**kwargs):
         "<{remaining}, {rate_fmt:>11}{postfix}]"
     )
     return tqdm(**kwargs)
-
-
-def scan_tqdm(
-    n: int,
-    print_rate: typing.Optional[int] = None,
-    message: typing.Optional[str] = None,
-) -> typing.Callable:
-    """
-    tqdm progress bar for a JAX scan
-
-    Parameters
-    ----------
-    n : int
-        Number of scan steps/iterations.
-    print_rate: int
-        Optional integer rate at which the progress bar will be updated,
-        by default the print rate will 1/20th of the total number of steps.
-    message : str
-        Optional string to prepend to tqdm progress bar.
-
-    Returns
-    -------
-    typing.Callable:
-        Progress bar wrapping function.
-    """
-
-    _update_progress_bar, close_tqdm = build_tqdm(n, print_rate, message)
-
-    def _scan_tqdm(func):
-        """Decorator that adds a tqdm progress bar to
-        `body_fun` used in `jax.lax.scan`. Note that
-        `body_fun` must either be looping over `jnp.arange(n)`,
-        or be looping over a tuple who's first element is `jnp.arange(n)`
-        This means that `iter_num` is the current iteration number
-        """
-
-        def wrapper_progress_bar(carry, x):
-            if type(x) is tuple:
-                iter_num, *_ = x
-            else:
-                iter_num = x
-            _update_progress_bar(iter_num)
-            result = func(carry, x)
-            return close_tqdm(result, iter_num)
-
-        return wrapper_progress_bar
-
-    return _scan_tqdm
-
-
-def loop_tqdm(
-    n: int,
-    print_rate: typing.Optional[int] = None,
-    message: typing.Optional[str] = None,
-) -> typing.Callable:
-    """
-    tqdm progress bar for a JAX fori_loop
-
-    Parameters
-    ----------
-    n : int
-        Number of iterations.
-    print_rate: int
-        Optional integer rate at which the progress bar will be updated,
-        by default the print rate will 1/20th of the total number of steps.
-    message : str
-        Optional string to prepend to tqdm progress bar.
-
-    Returns
-    -------
-    typing.Callable:
-        Progress bar wrapping function.
-    """
-
-    _update_progress_bar, close_tqdm = build_tqdm(n, print_rate, message)
-
-    def _loop_tqdm(func):
-        """
-        Decorator that adds a tqdm progress bar to `body_fun`
-        used in `jax.lax.fori_loop`.
-        """
-
-        def wrapper_progress_bar(i, val):
-            _update_progress_bar(i)
-            result = func(i, val)
-            return close_tqdm(result, i)
-
-        return wrapper_progress_bar
-
-    return _loop_tqdm
-
-
-def build_tqdm(
-    n: int,
-    print_rate: typing.Optional[int],
-    message: typing.Optional[str] = None,
-) -> typing.Tuple[typing.Callable, typing.Callable]:
-    """
-    Build the tqdm progress bar on the host
-    """
-
-    if message is None:
-        message = f"Running for {n:,} iterations"
-    tqdm_bars = {}
-
-    if print_rate is None:
-        if n > 20:
-            print_rate = int(n / 20)
-        else:
-            print_rate = 1
-    else:
-        if print_rate < 1:
-            raise ValueError(f"Print rate should be > 0 got {print_rate}")
-        elif print_rate > n:
-            raise ValueError(
-                "Print rate should be less than the "
[…]
