+++ hoi/utils/__init__.py
@@ -1,3 +1,3 @@
 from .logging import logger, set_log_level  # noqa
-from .progressbar import get_pbar, scan_tqdm, loop_tqdm  # noqa
+from .progressbar import get_pbar  # noqa
 from .stats import landscape, digitize, normalize, get_nbest_mult  # noqa
