+++ requirements.txt
@@ -4,4 +4,5 @@ numpy>=1.22
 scipy>=1.9
 scikit-learn
 pandas
+jax-tqdm
 tqdm
