### 2024-10-29 feb9eb7
Fix tests progressbar
