### 2024-10-29 d614edc
Rename plot_ by tuto_ + patch for sphinx gallery
