### 2024-10-29 brainets/hoi#64 @EtienneCmb
Indeed. We used the terminology `plot_` because sphinx-gallery uses this pattern to build the documentation. We changed it for `examples/it/tuto_core.py` in commit 29454b8
