+++ _downloads/016f1536d53dbce8b636f675e40ae7f4/plot_syn_red_mmi.py
@@ -201,6 +201,8 @@ plot_landscape(
     plt_kwargs=dict(cmap="turbo"),
 )
 
+plt.show()
+
 # %%
 # Plotting synergy
 
@@ -211,3 +213,5 @@ plot_landscape(
     undersampling=False,
     plt_kwargs=dict(cmap="turbo"),
 )
+
+plt.show()
