### 2024-10-29 brainets/hoi#65 @EtienneCmb
Thank you for reporting this issue.

October, Jax dropped some internal callback functions that we used to display the progress bar ([jax's changelog](https://github.com/jax-ml/jax/issues/20385)). I suspect that this error was due to the progress bar. We fixed this issue with commits 38cc6bc and feb9eb7. Tests are now [passing](https://github.com/brainets/hoi/actions/runs/11572708095) for python>=3.8.

Could you please fetch the latest commits, update jax with `pip install -U jax` and running this example again?
