### 2024-10-29 brainets/hoi#67 @EtienneCmb
See my answer to #65
