### 2024-10-29 brainets/hoi#68 @EtienneCmb
Indeed, we used `plot_jax` for the same reason as #64. We renamed the file for `examples/miscellaneous/plot_jax.py`.

Thanks !
