### 2024-10-29 brainets/hoi#70 @EtienneCmb
Hi @pitmonticone,

Thank you for reporting this warning. This is probably due to the use of LateX with matplotlib. Unfortunately, I can't reproduce this behavior on my machine (Linux). And the [online](https://brainets.github.io/hoi/auto_examples/tutorials/plot_ml_vs_it.html#sphx-glr-auto-examples-tutorials-plot-ml-vs-it-py) example doesn't show this warning either.

Could you please try to update matplotlib `pip install -U matplotlib`? What's your OS?
