### 2024-10-29 brainets/hoi#60 @pitmonticone
I consider this issue resolved.
