### 2024-10-29 brainets/hoi#72 @EtienneCmb
Hi @ClaudMor,

Indeed, the code failed because there's only a single triplet with a dataset composed of three variables. This issue should be solved with d2c98bb.

Could you please try to fetch the latest version of hoi and try again?
