### 2024-10-29 brainets/hoi#74 @EtienneCmb
Hi @ClaudMor,

I'm able to reproduce the issue. Regarding the performance claims about [higher-order](https://brainets.github.io/hoi/jax.html#computing-higher-order-interactions-on-large-multiplets), it still works, but for a reason, it doesn't work anymore with the entropy. I'll let you know once I figure it out ! Thanks !
