### 2024-10-29 brainets/hoi#70 @pitmonticone
I have updated `matplotlib`, but the SyntaxWarning remains.

I'm running it on macOS.
