### 2024-10-29 brainets/hoi#65 @pitmonticone
Fetched the latest commits and updated jax.

Here is the output:

```console
Copnorm and demean the data
Get list of multiplets
   index  order       hoi           multiplet
0     72      5  0.182472     [0, 1, 2, 3, 6]
1     35      4  0.181906        [0, 1, 2, 3]
2      0      3  0.181892           [0, 1, 2]
3     37      4  0.181686        [0, 1, 2, 5]
4     38      4  0.181507        [0, 1, 2, 6]
5     84      5 -0.447083     [0, 3, 4, 5, 6]
6     69      4 -0.447269        [3, 4, 5, 6]
7     89      5 -0.453236     [1, 3, 4, 5, 6]
8     97      6 -0.453758  [1, 2, 3, 4, 5, 6]
9     95      6 -0.454479  [0, 1, 3, 4, 5, 6]
Compute entropy with gc
    Copnorm and demean the data
Get list of multiplets
Get list of multiplets
Traceback (most recent call last):
  File "/Users/pietromonticone/GitHub/hoi/examples/metrics/plot_infotopo.py", line 99, in <module>
    hoi = model.fit(minsize=3, method="gc")
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/pietromonticone/GitHub/hoi/.venv/lib/python3.12/site-packages/hoi/metrics/infotopo.py", line 154, in fit
    _, hoi = jax.lax.scan(
             ^^^^^^^^^^^^^
  File "/Users/pietromonticone/GitHub/hoi/.venv/lib/python3.12/site-packages/hoi/utils/progressbar.py", line 62, in wrapper_progress_bar
    _update_progress_bar(iter_num)
  File "/Users/pietromonticone/GitHub/hoi/.venv/lib/python3.12/site-packages/hoi/utils/progressbar.py", line 151, in _update_progress_bar
    _ = jax.jax.lax.cond(
        ^^^^^^^^^^^^^^^^^
  File "/Users/pietromonticone/GitHub/hoi/.venv/lib/python3.12/site-packages/hoi/utils/progressbar.py", line 153, in <lambda>
    lambda _: host_callback.id_tap(
              ^^^^^^^^^^^^^^^^^^^^^
  File "/Users/pietromonticone/GitHub/hoi/.venv/lib/python3.12/site-packages/jax/experimental/host_callback.py", line 27, in call
    raise NotImplementedError(
NotImplementedError: jax.experimental.host_callback has been deprecated since March 2024 and is now no longer supported. See https://github.com/jax-ml/jax/issues/20385
--------------------
For simplicity, JAX has removed its internal frames from the traceback of the following exception. Set JAX_TRACEBACK_FILTERING=off to include these.
```
