### 2024-10-29 brainets/hoi#65 @EtienneCmb
I believe there's an issue with your update because the latest version of hoi doesn't have the [progress bar](https://github.com/brainets/hoi/blob/main/hoi%2Futils%2Fprogressbar.py) anymore. Now we rely on the package jax-tqdm. Have you installed hoi using pip or python setup.py?
