### 2024-10-29 brainets/hoi#65 @pitmonticone
I have installed in my local `.venv` using `pip install -U hoi`
