### 2024-10-30 brainets/hoi#65 @EtienneCmb
Hi @pitmonticone , the new version is on [pypi](https://pypi.org/project/hoi/)
