### 2024-10-30 brainets/hoi#65 @EtienneCmb
> Yes, now it works fine:
>
> ```console
>   Copnorm and demean the data
> Get list of multiplets
>    index  order       hoi           multiplet
> 0     36      4  0.203451        [0, 1, 2, 4]
> 1     73      5  0.203430     [0, 1, 2, 4, 5]
> 2     37      4  0.203373        [0, 1, 2, 5]
> 3      0      3  0.203372           [0, 1, 2]
> 4     72      5  0.202888     [0, 1, 2, 3, 6]
> 5     96      6 -0.420208  [0, 2, 3, 4, 5, 6]
> 6     84      5 -0.420437     [0, 3, 4, 5, 6]
> 7     95      6 -0.420715  [0, 1, 3, 4, 5, 6]
> 8     97      6 -0.421467  [1, 2, 3, 4, 5, 6]
> 9     89      5 -0.422012     [1, 3, 4, 5, 6]
> Compute entropy with gc
>     Copnorm and demean the data
> Get list of multiplets
> Get list of multiplets
> Mutual information: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 99/99 [00:00<00:00, 75980.99it/s]
>    index  order       hoi     multiplet
> 0      0      3  0.203372     [0, 1, 2]
> 1     69      4  0.094784  [3, 4, 5, 6]
> 2      4      3  0.000405     [0, 1, 6]
> 3     62      4  0.000385  [1, 3, 4, 6]
> 4     52      4  0.000381  [0, 3, 4, 6]
> 5     47      4 -0.000550  [0, 2, 3, 6]
> 6      8      3 -0.001450     [0, 2, 6]
> 7     33      3 -0.057861     [3, 5, 6]
> 8     31      3 -0.060821     [3, 4, 5]
> 9     32      3 -0.110925     [3, 4, 6]
> ```
>
> Sorry for the pedantry, but during review I need to test the user experience.
>
> I will close all the similar issues.

Great ! No problem, thank you for reporting this !
