### 2024-10-30 brainets/hoi#67 @pitmonticone
Same here:

```console
Compute entropy with gc
    Copnorm and demean the data
Get list of multiplets
Get list of multiplets
Mutual information: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 42/42 [00:00<00:00, 44824.62it/s]
   index  order       hoi     multiplet
0      0      3  0.196291     [0, 1, 2]
1     10      3  0.004352     [1, 2, 3]
2     13      3  0.004202     [1, 3, 4]
3     30      4  0.001595  [1, 2, 3, 4]
4      2      3  0.001399     [0, 1, 4]
5     27      4 -0.001132  [0, 2, 3, 5]
6     14      3 -0.001670     [1, 3, 5]
7      8      3 -0.001762     [0, 3, 5]
8     24      4 -0.002510  [0, 1, 3, 5]
9     19      3 -0.222632     [3, 4, 5]
```

I consider this issue resolved.
