### 2024-10-30 brainets/hoi#70 @pitmonticone
Please consider using raw strings:

```python
plt.ylabel(r"$\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
```

or double backslash:

```python
plt.ylabel("$\\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
```
