### 2024-10-30 brainets/hoi#70 @EtienneCmb
At the top of the script, there's the line `plt.style.use('ggplot')`. Could you please try to comment this line and run the script again?
