### 2024-10-30 brainets/hoi#70 @pitmonticone
Same output.
