### 2024-10-30 c8a53d9
Raw strings in the ML vs. IT example
