+++ examples/tutorials/plot_ml_vs_it.py
@@ -185,11 +185,11 @@ fig, axs = plt.subplots(2, 1, figsize=(8, 12), sharex=True, sharey=True)
 axs = np.ravel(axs)
 plt.sca(axs[0])
 plt.plot(it(x_red, y_red).squeeze())
-plt.ylabel("$\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
+plt.ylabel(r"$\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
 plt.title("Redundant interactions", fontweight="bold")
 plt.sca(axs[1])
 plt.plot(it(x_syn, y_syn).squeeze())
-plt.ylabel("$\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
+plt.ylabel(r"$\partial \Omega_{[X_{1}, ..., X_{N}]}$ [Bits]")
 plt.title("Synergistic interactions", fontweight="bold")
 plt.show()
 
