### 2024-10-30 brainets/hoi#72 @EtienneCmb
You can use `pip install -U hoi` to get the patch
