+++ hoi/__init__.py
@@ -1,3 +1,3 @@
 from hoi import core, metrics, utils, simulation  # noqa
 
-__version__ = "0.0.2"
+__version__ = "0.0.3"
