+++ setup.py
@@ -4,7 +4,7 @@
 import os
 from setuptools import setup, find_packages
 
-__version__ = "0.0.2"
+__version__ = "0.0.3"
 NAME = "hoi"
 AUTHOR = "BraiNets"
 MAINTAINER = "Etienne Combrisson"
