### 2024-10-30 brainets/hoi#72 @pitmonticone
I have just run

```python
from hoi.simulation import simulate_hoi_gauss

data = simulate_hoi_gauss(n_samples=1000, triplet_character='synergy')

# import the O-information
from hoi.metrics import Oinfo

# define the model
model = Oinfo(data)

# compute hoi for multiplets with a minimum size of 3 and maximum size of 3
# using the Gaussian Copula entropy
hoi = model.fit(minsize=3, maxsize=3, method="gc")

from hoi.plot import plot_landscape
from hoi.utils import get_nbest_mult

# plot the landscape
plot_landscape(hoi, model=model)

# print the summary table
print(get_nbest_mult(hoi, model=model))
```

and the output is error free:

```console
    Copnorm and demean the data
Get list of multiplets
   index  order       hoi  multiplet
0      0      3 -0.846935  [0, 1, 2]
```
