### 2024-10-30 brainets/hoi#70 @EtienneCmb
> Could you please publish the new version on pypi?

Version [0.0.3](https://pypi.org/project/hoi/) is on pypi
