### 2024-10-30 brainets/hoi#72 @pitmonticone
I consider this issue resolved. However, I will await confirmation from @ClaudMor for verification.
