### 2024-10-30 brainets/hoi#74 @EtienneCmb
Hi @ClaudMor,

I fixed the issue. I believe it was coming from a bad way of timing computations. Here's the revised code :

```python
import numpy as np
import jax
import jax.numpy as jnp
import timeit

from hoi.core import get_entropy

import matplotlib.pyplot as plt

plt.style.use("ggplot")

n = 5
n_repeat= 5

entropy = jax.jit(jax.vmap(get_entropy(method="gc"), in_axes=(0,)))
entropy(np.random.rand(2, 2, 10))  # dry run

n_samples = np.linspace(10, 10e2, n).astype(int)
n_features = np.linspace(1, 10, n).astype(int)
n_variables = np.linspace(1, 10e2, n).astype(int)

data_size, timings_gpu, timings_cpu = [], [], []
for n_s, n_f, n_v in zip(n_samples, n_features, n_variables):
    x = np.random.rand(n_v, n_f, n_s)
    x = jnp.asarray(x)

    with jax.default_device(jax.devices("cpu")[0]):
        result_cpu = timeit.timeit('entropy(x).block_until_ready()', number=n_repeat, globals=globals())
        timings_cpu.append(result_cpu / n_repeat)

    with jax.default_device(jax.devices("gpu")[0]):
        result_gpu = timeit.timeit('entropy(x).block_until_ready()', number=n_repeat, globals=globals())
        timings_gpu.append(result_gpu / n_repeat)

    data_size.append(n_s * n_f * n_v)

plt.plot(data_size, timings_cpu, label="CPU")
plt.plot(data_size, timings_gpu, label="GPU")
plt.xlabel("Data size")
plt.ylabel("Time (s)")
plt.title("CPU vs. GPU for computing entropy", fontweight="bold")
plt.legend()
```
Very similar results on Google colab and Kaggle :
![image](https://github.com/user-attachments/assets/3fcd260b-b66d-40bb-9957-d6e2742fd4cc)

I let you try this code snippet and if you've similar results, I'll update the online documentation accordingly
