### 2024-10-30 5110377
Added section "Related packages" to the paper
