### 2024-10-30 brainets/hoi#73 @EtienneCmb
Hi @pitmonticone,

I added the "Related packages" section to the draft (see commit 5110377 and the [pdf](https://github.com/openjournals/joss-reviews/issues/7360#issuecomment-2446939409) produced). The last remaining point about the missing DOI has been fixed yesterday
