### 2024-10-30 brainets/hoi#73 @pitmonticone
Ok, I consider this issue to be resolved.
