### 2024-10-31 brainets/hoi#74 @ClaudMor
Hi @EtienneCmb ,

I confirm I now get similar results.
