### 2024-10-31 #7360 @pitmonticone
> I have finished my review:
>
> - [Review Issues](https://github.com/brainets/hoi/issues?q=is%3Aissue+author%3Apitmonticone)
>
> - [Review PRs](https://github.com/brainets/hoi/pulls?q=is%3Apr+author%3Apitmonticone+)

All the issues I opened have been resolved and all the PRs I opened have been merged.

I recommend this paper for acceptance.
