### 2024-10-31 brainets/hoi#74 @EtienneCmb
Great ! I'll update the doc soon
