### 2024-10-31 #7360 @ClaudMor
Hi @jbytecode,

The authors have satisfied all my issues and I believe the package represents a valid contribution. I invite the authors to update the performance claims as agreed [here](https://github.com/brainets/hoi/issues/74) and recommend acceptance.
