### 2024-10-31 fdb1301
Fix timing computations on C/GPU for entropy
