### 2024-10-31 brainets/hoi#74 @EtienneCmb
Hi @ClaudMor, I fixed the doc as we discussed in commit fdb1301
