### 2024-11-05 #7360 @ClaudMor
Hi @jbytecode and @EtienneCmb,

I confirm I now get the expected plots.
