+++ setup.py
@@ -4,7 +4,7 @@
 import os
 from setuptools import setup, find_packages
 
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 NAME = "hoi"
 AUTHOR = "BraiNets"
 MAINTAINER = "Etienne Combrisson"
