### 2024-11-06 #7360 @EtienneCmb
Hi @jbytecode,

- [x] Authors and affiliations are correct
- [x] We released the version [0.0.5](https://github.com/brainets/hoi/releases/tag/v0.0.5)
- [x] Zenodo's DOI : [10.5281/zenodo.14046562](https://doi.org/10.5281/zenodo.14046562)
- [x] I checked title, authors and license file, all of them match those in the JOSS paper
