### 2024-11-07 #7360 @jbytecode
@EtienneCmb - I've just sent a pull request. This PR includes some minor changes in both the manuscript and the bibliography. Could you please review the PR and merge it? Thank you in advance.

https://github.com/brainets/hoi/pull/75
