### 2024-11-07 #7360 @EtienneCmb
Hi @jbytecode, thanks for the PR, I merged it !
