### 2024-11-07 #7360 @jbytecode
@EtienneCmb - The paper and the bibliography now look good to me. Please perform one more proof-reading of the paper. Please tell me if there is any issue. Thank you in advance.
