### 2024-11-08 #7360 @EtienneCmb
@jbytecode the draft looks correct indeed 👍
