### 2024-11-08 #7360 @EtienneCmb
@jbytecode your PR has been merged, thank you
