### 2024-11-09 #7360 @crvernon
@EtienneCmb - great work on this one.  I just need you to correct the following before I move to accept this for publication:

In the paper:

- [ ] LINES 32-33:  Note the extra "2010" at the end of the in text citations.  I believe this is because you reference `@williams:2010` twice in your paper.md file here.  Please fix this.

Let me know when this has been fixed.  Thanks!
