+++ paper/paper.md
@@ -56,7 +56,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Recent research studying higher-order interactions with information theoretic measures provides new angles and valuable insights in different fields, such as neuroscience [@gatica:2021; @herzog:2022; @combrisson:2024; @luppi:2022; @baudot:2019], music [@rosas:2019], economics [@scagliarini:2023], and psychology [@marinazzo:2022]. Information theory allows investigating higher-order interactions using a rich set of metrics that provide interpretable values of the statistical interdependency among multivariate data [@williams:2010; @mediano:2021; @barrett:2015; @rosas:2019; @scagliarini:2023; @williams:2010].
+Recent research studying higher-order interactions with information theoretic measures provides new angles and valuable insights in different fields, such as neuroscience [@gatica:2021; @herzog:2022; @combrisson:2024; @luppi:2022; @baudot:2019], music [@rosas:2019], economics [@scagliarini:2023], and psychology [@marinazzo:2022]. Information theory allows investigating higher-order interactions using a rich set of metrics that provide interpretable values of the statistical interdependency among multivariate data [@williams:2010; @mediano:2021; @barrett:2015; @rosas:2019; @scagliarini:2023].
 
 Despite the relevance of studying higher-order interactions across various fields, there is currently no toolkit that compiles the latest approaches and offers user-friendly functions for calculating higher-order information metrics. Computing higher-order information presents two main challenges. First, these metrics rely on entropy and mutual information, whose estimation must be adapted to different types of data [@madukaife:2024; @czyz:2024]. Second, the computational complexity increases exponentially as the number of variables and interaction orders grows. For example, a dataset with 100 variables, has approximately $1.6 \times 10^5$ possible triplets, $4 \times 10^6$ quadruplets, and $7 \times 10^7$ quintuplets. Therefore, an efficient implementation, scalable on modern hardware is required.
 
