### 2024-11-12 #7360 @EtienneCmb
Hi @crvernon, thank you for pointing this mistake. Indeed it was coming from a double citation. This should be fixed by [34fb954](https://github.com/brainets/hoi/commit/34fb954d24e2504b5f5bec9d17a8285c78d0810c)
